code replacement and new comments added
